util: Move common type conversion helpers into util section
These helpers are useful not just in the core code. Allow plugins
to use these too.
