+++ util/meson.build
@@ -6,4 +6,5 @@ sources += [
   'util/json.c',
   'util/suffix.c',
   'util/base64.c',
+  'util/types.c',
 ]
